refactor: Move screens for groups and openings into nested packages.
