+++ app/src/main/java/no/usn/mob3000/ui/CBViewModel.kt
@@ -3,7 +3,7 @@ package no.usn.mob3000.ui
 import androidx.compose.runtime.State
 import androidx.compose.runtime.mutableStateOf
 import androidx.lifecycle.ViewModel
-import no.usn.mob3000.ui.screens.chess.train.Opening
+import no.usn.mob3000.ui.screens.chess.train.opening.Opening
 
 /**
  * ViewModel to save state where necessary.
