+++ app/src/main/java/no/usn/mob3000/ui/screens/chess/train/opening/CreateOpeningScreen.kt
@@ -1,4 +1,4 @@
-package no.usn.mob3000.ui.screens.chess.train
+package no.usn.mob3000.ui.screens.chess.train.opening
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
@@ -7,19 +7,17 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.res.stringResource
-import no.usn.mob3000.R
 import no.usn.mob3000.Viewport
 
 @Composable
-fun CreateGroupScreen() {
+fun CreateOpeningScreen() {
     Viewport { innerPadding ->
         Box(
             modifier = Modifier.fillMaxSize()
                                .padding(innerPadding),
             contentAlignment = Alignment.Center
         ) {
-            Text("Create group screen. Stub.")
+            Text("Create opening screen. Stub.")
         }
     }
 }
